chore: using team fours sorting module
