--- bosco/sorting.py
@@ -1,358 +1,0 @@
-"""Sorting algorithms for lists that contain integer values."""
-
-from random import randint
-from typing import List
-
-
-def bubble_sort(array: List[int]) -> List[int]:
-    """Sort an input list called array using bubble sort."""
-    n = len(array)
-
-    for i in range(n):
-        # Create a flag that will allow the function to
-        # terminate early if there's nothing left to sort
-        already_sorted = True
-
-        # Start looking at each item of the list one by one,
-        # comparing it with its adjacent value. With each
-        # iteration, the portion of the array that you look at
-        # shrinks because the remaining items have already been
-        # sorted.
-        for j in range(n - i - 1):
-            if array[j] > array[j + 1]:
-                # If the item you're looking at is greater than its
-                # adjacent value, then swap them
-                array[j], array[j + 1] = array[j + 1], array[j]
-
-                # Since you had to swap two elements,
-                # set the `already_sorted` flag to `False` so the
-                # algorithm doesn't finish prematurely
-                already_sorted = False
-
-        # If there were no swaps during the last iteration,
-        # the array is already sorted, and you can terminate
-        if already_sorted:
-            break
-
-    return array
-
-
-def insertion_sort(array: List[int]) -> List[int]:
-    """Run an insertion sort on the provided array."""
-    # Loop from the second element of the array until
-    # the last element
-    for i in range(1, len(array)):
-        # This is the element we want to position in its
-        # correct place
-        key_item = array[i]
-
-        # Initialize the variable that will be used to
-        # find the correct position of the element referenced
-        # by `key_item`
-        j = i - 1
-
-        # Run through the list of items (the left
-        # portion of the array) and find the correct position
-        # of the element referenced by `key_item`. Do this only
-        # if `key_item` is smaller than its adjacent values.
-        while j >= 0 and array[j] > key_item:
-            # Shift the value one position to the left
-            # and reposition j to point to the next element
-            # (from right to left)
-            array[j + 1] = array[j]
-            j -= 1
-
-        # When you finish shifting the elements, you can position
-        # `key_item` in its correct location
-        array[j + 1] = key_item
-
-    return array
-
-
-def merge(left: List[int], right: List[int]) -> List[int]:
-    """Define a convenience method that supports the merging of lists."""
-    # If the first array is empty, then nothing needs
-    # to be merged, and you can return the second array as the result
-    if len(left) == 0:
-        return right
-
-    # If the second array is empty, then nothing needs
-    # to be merged, and you can return the first array as the result
-    if len(right) == 0:
-        return left
-
-    result: List[int] = []
-    index_left = index_right = 0
-
-    # Now go through both arrays until all the elements
-    # make it into the resultant array
-    while len(result) < len(left) + len(right):
-        # The elements need to be sorted to add them to the
-        # resultant array, so you need to decide whether to get
-        # the next element from the first or the second array
-        if left[index_left] <= right[index_right]:
-            result.append(left[index_left])
-            index_left += 1
-        else:
-            result.append(right[index_right])
-            index_right += 1
-
-        # If you reach the end of either array, then you can
-        # add the remaining elements from the other array to
-        # the result and break the loop
-        if index_right == len(right):
-            result += left[index_left:]
-            break
-
-        if index_left == len(left):
-            result += right[index_right:]
-            break
-
-    return result
-
-
-def merge_sort(array: List[int]) -> List[int]:
-    """Sort the provided list called array with the merge sort algorithm."""
-    # If the input array contains fewer than two elements,
-    # then return it as the result of the function
-    if len(array) < 2:
-        return array
-
-    midpoint = len(array) // 2
-
-    # Sort the array by recursively splitting the input
-    # into two equal halves, sorting each half and merging them
-    # together into the final result
-    return merge(
-        left=merge_sort(array[:midpoint]), right=merge_sort(array[midpoint:])
-    )
-
-
-def quick_sort(array: List[int]) -> List[int]:
-    """Sort the provided list called array with the quick sort algorithm."""
-    # If the input array contains fewer than two elements,
-    # then return it as the result of the function
-    if len(array) < 2:
-        return array
-
-    low, same, high = [], [], []
-
-    # Select your `pivot` element randomly
-    pivot = array[randint(0, len(array) - 1)]
-
-    for item in array:
-        # Elements that are smaller than the `pivot` go to
-        # the `low` list. Elements that are larger than
-        # `pivot` go to the `high` list. Elements that are
-        # equal to `pivot` go to the `same` list.
-        if item < pivot:
-            low.append(item)
-        elif item == pivot:
-            same.append(item)
-        elif item > pivot:
-            high.append(item)
-
-    # The final result combines the sorted `low` list
-    # with the `same` list and the sorted `high` list
-    return quick_sort(low) + same + quick_sort(high)
-
-
-def insertion_sort_tim(array: List[int], left: int = 0, right=None):
-    """Use an internal sorting algorithm for the timsort algorithm."""
-    if right is None:
-        right = len(array) - 1
-
-    # Loop from the element indicated by
-    # `left` until the element indicated by `right`
-    for i in range(left + 1, right + 1):
-        # This is the element we want to position in its
-        # correct place
-        key_item = array[i]
-
-        # Initialize the variable that will be used to
-        # find the correct position of the element referenced
-        # by `key_item`
-        j = i - 1
-
-        # Run through the list of items (the left
-        # portion of the array) and find the correct position
-        # of the element referenced by `key_item`. Do this only
-        # if the `key_item` is smaller than its adjacent values.
-        while j >= left and array[j] > key_item:
-            # Shift the value one position to the left
-            # and reposition `j` to point to the next element
-            # (from right to left)
-            array[j + 1] = array[j]
-            j -= 1
-
-        # When you finish shifting the elements, position
-        # the `key_item` in its correct location
-        array[j + 1] = key_item
-
-    return array
-
-
-def tim_sort(array: List[int]) -> List[int]:
-    """Sort the list called array with the tim sort algorithm using a special insertion sort."""
-    min_run = 32
-    n = len(array)
-
-    # Start by slicing and sorting small portions of the
-    # input array. The size of these slices is defined by
-    # your `min_run` size.
-    for i in range(0, n, min_run):
-        insertion_sort_tim(array, i, min((i + min_run - 1), n - 1))
-
-    # Now you can start merging the sorted slices.
-    # Start from `min_run`, doubling the size on
-    # each iteration until you surpass the length of
-    # the array.
-    size = min_run
-    while size < n:
-        # Determine the arrays that will
-        # be merged together
-        for start in range(0, n, size * 2):
-            # Compute the `midpoint` (where the first array ends
-            # and the second starts) and the `endpoint` (where
-            # the second array ends)
-            midpoint = start + size - 1
-            end = min((start + size * 2 - 1), (n - 1))
-
-            # Merge the two subarrays.
-            # The `left` array should go from `start` to
-            # `midpoint + 1`, while the `right` array should
-            # go from `midpoint + 1` to `end + 1`.
-            merged_array = merge(
-                left=array[start : midpoint + 1],
-                right=array[midpoint + 1 : end + 1],
-            )
-
-            # Finally, put the merged array back into
-            # your array
-            array[start : start + len(merged_array)] = merged_array
-
-        # Each iteration should double the size of your arrays
-        size *= 2
-
-    return array
-
-
-# Selection Sort algorithm in Python
-def selection_sort(array):
-    for s in range(len(array)):
-        min_idx = s
-
-        for i in range(s + 1, len(array)):
-            # For sorting in descending order
-            # for minimum element in each loop
-            if array[i] < array[min_idx]:
-                min_idx = i
-
-        # Arranging min at the correct position
-        (array[s], array[min_idx]) = (array[min_idx], array[s])
-
-
-def heapify(array, n, i):
-    largest = i
-    l = 2 * i + 1
-    r = 2 * i + 2
-
-    if l < n and array[i] < array[l]:
-        largest = l
-    if r < n and array[largest] < array[r]:
-        largest = r
-
-    if largest != i:
-        array[i], array[largest] = array[largest], array[i]
-        heapify(array, n, largest)
-
-
-def heap_sort(array):
-    n = len(array)
-    for i in range(n // 2, -1, -1):
-        heapify(array, n, i)
-    for i in range(n - 1, 0, -1):
-        array[i], array[0] = array[0], array[i]
-        heapify(array, i, 0)
-    return array
-
-
-def shell_sort(array):
-    n = len(array)
-    interval = n // 2
-    while interval > 0:
-        for i in range(interval, n):
-            temp = array[i]
-            j = i
-            while j >= interval and array[j - interval] > temp:
-                array[j] = array[j - interval]
-                j -= interval
-
-            array[j] = temp
-        interval //= 2
-    return array
-
-
-# Radix sort in Python
-
-
-# Using counting sort to sort the elements in the basis of significant places
-def countingSort_radix(array, place):
-    size = len(array)
-    output = [0] * size
-    count = [0] * 10
-
-    # Calculate count of elements
-    for i in range(0, size):
-        index = array[i] // place
-        count[index % 10] += 1
-
-    # Calculate cumulative count
-    for i in range(1, 10):
-        count[i] += count[i - 1]
-
-    # Place the elements in sorted order
-    i = size - 1
-    while i >= 0:
-        index = array[i] // place
-        output[count[index % 10] - 1] = array[i]
-        count[index % 10] -= 1
-        i -= 1
-
-    for i in range(0, size):
-        array[i] = output[i]
-
-
-# Main function to implement radix sort
-def radix_sort(array):
-    # Get maximum element
-    max_element = max(array)
-
-    # Apply counting sort to sort elements based on place value.
-    place = 1
-    while max_element // place > 0:
-        countingSort_radix(array, place)
-        place *= 10
-    return array
-
-
-def bucket_sort(arr):
-    n = len(arr)
-    buckets = [[] for _ in range(n)]
-
-    # Put array elements in different buckets
-    for num in arr:
-        bi = min(int(n * num), n - 1)
-        buckets[bi].append(num)
-
-    # Sort individual buckets using insertion sort
-    for bucket in buckets:
-        insertion_sort(bucket)
-
-    # Concatenate all buckets into arr[]
-    index = 0
-    for bucket in buckets:
-        for num in bucket:
-            arr[index] = num
-            index += 1
-    return arr
